Update opening validation in PHAST
