+++ src/app/calculator/furnaces/opening/opening-form/opening-form.component.html
@@ -189,9 +189,7 @@
       </div>
   
       <div class="form-group">
-        <label class="small" for="{{'viewFactor'+idString}}" aria-describedby="viewFactorHelp">View Factor
-          <!-- <a id="{{'viewFactor_'+lossIndex}}" class="form-text small click-link" (click)="calculateViewFactor()">Calculate View Factor</a> -->
-        </label>
+        <label class="small" for="{{'viewFactor'+idString}}" aria-describedby="viewFactorHelp">View Factor</label>
   
         <input  name="{{'viewFactor_'+lossIndex}}" type="number" step="0.1" min="0" max="1"
           class="form-control" formControlName="viewFactor" id="{{'viewFactor'+idString}}" onfocus="this.select();"
