+++ src/app/phast/losses/losses-tabs/opening-tab/opening-tab.component.ts
@@ -2,10 +2,11 @@ import { Component, OnInit, Input, ChangeDetectorRef } from '@angular/core';
 import { LossesService } from '../../losses.service';
 import { PHAST } from '../../../../shared/models/phast/phast';
 import { FormGroup } from '@angular/forms';
-import { OpeningLossesService, OpeningLossWarnings } from '../../opening-losses/opening-losses.service';
 import { OpeningLossesCompareService } from '../../opening-losses/opening-losses-compare.service';
 import { OpeningLoss } from '../../../../shared/models/phast/losses/openingLoss';
 import { Subscription } from 'rxjs';
+import { OpeningLossWarnings, OpeningService } from '../../../../calculator/furnaces/opening/opening.service';
+import { OpeningFormService } from '../../../../calculator/furnaces/opening/opening-form.service';
 
 @Component({
   selector: 'app-opening-tab',
@@ -27,7 +28,10 @@ export class OpeningTabComponent implements OnInit {
   isDifferent: boolean;
   badgeClass: Array<string> = [];
   lossSubscription: Subscription;
-  constructor(private lossesService: LossesService, private openingLossesService: OpeningLossesService, private openingLossesCompareService: OpeningLossesCompareService, private cd: ChangeDetectorRef) { }
+  constructor(private lossesService: LossesService, 
+              private openingLossesService: OpeningService,
+              private openingFormService: OpeningFormService, 
+              private openingLossesCompareService: OpeningLossesCompareService, private cd: ChangeDetectorRef) { }
 
   ngOnInit() {
     this.setNumLosses();
@@ -98,7 +102,7 @@ export class OpeningTabComponent implements OnInit {
 
 
   checkLossValid(loss: OpeningLoss) {
-      let tmpForm: FormGroup = this.openingLossesService.getFormFromLoss(loss);
+      let tmpForm: FormGroup = this.openingFormService.getFormFromLoss(loss);
       if (tmpForm.status === 'VALID') {
         return true;
       } else {
