+++ src/app/phast/losses/opening-losses/opening-losses.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
-import { OpeningLossesService } from './opening-losses.service';
 import { OpeningLossesCompareService } from './opening-losses-compare.service';
 import { OpeningLossesComponent } from './opening-losses.component';
 import { OpeningLossesFormComponent } from './opening-losses-form/opening-losses-form.component';
@@ -19,7 +18,6 @@ import { SharedPipesModule } from '../../../shared/shared-pipes/shared-pipes.mod
   ],
   providers: [
     OpeningLossesCompareService,
-    OpeningLossesService
   ],
   exports: [
     OpeningLossesComponent
