--- src/app/phast/losses/opening-losses/opening-losses.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { OpeningLossesService } from './opening-losses.service';
-
-describe('OpeningLossesService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [OpeningLossesService]
-    });
-  });
-
-  it('should ...', inject([OpeningLossesService], (service: OpeningLossesService) => {
-    expect(service).toBeTruthy();
-  }));
-});
